BL-15 move pages components to pages
