+++ src/components/index.ts
@@ -1,11 +1,6 @@
 // This file is used to export all components from a single file
-export { default as Landing } from "./Landing";
 export { default as Sidebar } from "./Sidebar";
 export { default as NavigationSidebar } from "./NavigationSidebar";
 export { default as SearchBar } from "./SearchBar";
-export { default as Favourites } from "./Favourites";
-export { default as Bestsellers } from "./Bestsellers";
 export { default as ListItem } from "./ListItem";
 export { default as Rating } from "./Rating";
-export { default as UpdateBook } from "./UpdateBook";
-export { default as Settings } from "./Settings";
