+++ src/pages/Favourites/Favourites.styles.ts
