+++ src/pages/UpdateBook/UpdateBook.interfaces.ts
@@ -1,4 +1,4 @@
-import { Book } from "src/components/Bestsellers/Bestsellers.interfaces";
+import { Book } from "src/pages/Bestsellers/Bestsellers.interfaces";
 
 export interface UpdateBookProps {
   favourites: Book[];
