+++ src/pages/index.ts
@@ -1,0 +1,6 @@
+// This file is used to export all pages components from a single file
+export { default as Landing } from "./Landing";
+export { default as Favourites } from "./Favourites";
+export { default as Bestsellers } from "./Bestsellers";
+export { default as UpdateBook } from "./UpdateBook";
+export { default as Settings } from "./Settings";
